+++ packages/components/src/components/DataAnalytics.tsx
@@ -6,7 +6,7 @@ import { variables } from '../config';
 import { Button } from './buttons/Button/Button';
 import { CollapsibleBox } from './CollapsibleBox';
 import { CollapsibleCard } from './CollapsibleCard';
-import { Switch } from './form/Switch';
+import { Switch } from './form/Switch/Switch';
 import { H1 } from './typography/Heading/Heading';
 
 const Wrapper = styled.div`
