+++ packages/components/src/index.ts
@@ -14,13 +14,13 @@ export * from './components/buttons/PinButton/PinButton';
 export * from './components/buttons/TooltipButton/TooltipButton';
 export * from './components/Flag/Flag';
 export * from './components/Dropdown/Dropdown';
-export * from './components/form/Input';
-export * from './components/form/Textarea';
-export * from './components/form/Select';
-export * from './components/form/Checkbox';
-export * from './components/form/RadioButton';
+export * from './components/form/Input/Input';
+export * from './components/form/Textarea/Textarea';
+export * from './components/form/Select/Select';
+export * from './components/form/Checkbox/Checkbox';
+export * from './components/form/RadioButton/RadioButton';
 export * from './components/form/Range/Range';
-export * from './components/form/Switch';
+export * from './components/form/Switch/Switch';
 export * from './components/form/InputStyles';
 export * from './components/Icon/Icon';
 export * from './components/typography/Heading/Heading';
@@ -42,7 +42,7 @@ export * from './components/loaders/LoadingContent';
 export * from './components/Tooltip/Tooltip';
 export * from './components/Timerange/Timerange';
 export * from './components/Truncate/Truncate';
-export * from './components/form/SelectBar';
+export * from './components/form/SelectBar/SelectBar';
 export * from './components/HoverAnimation';
 export * from './components/Fade';
 export * from './components/Image/Image';
